% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Funcion que da la bienvenida al equipo de KOE}
\usage{
hello_world(x)
}
\arguments{
\item{x}{string con nombre del integrante del equipo}
}
\value{
un vector con un string de saludo
}
\description{
Funcion que da la bienvenida al equipo de KOE
}
\examples{
hello_world(x = "Bastian")
}
